+++ example/lib/main.dart
@@ -1,0 +1,6 @@
+import 'package:example/app/app.dart';
+import 'package:flutter/material.dart';
+
+void main() {
+  runApp(const AlchemistExampleApp());
+}