+++ test/src/golden_test_runner_test.dart
@@ -18,7 +18,7 @@ void main() {
     registerFallbackValue(MockWidgetTester());
     registerFallbackValue(const BoxConstraints());
     registerFallbackValue(ThemeData.light());
-    registerFallbackValue(const SizedBox.square());
+    registerFallbackValue(const SizedBox());
     registerFallbackValue(find.byType(Widget));
   });
 
@@ -44,48 +44,19 @@ void main() {
   group('GoldenTestRunner', () {
     const goldenTestRunner = FlutterGoldenTestRunner();
     late MockAdapter adapter;
-    var cleanupCalled = false;
-    var interactionCalled = false;
-    var goldenFileExpectationCalled = false;
-    var matcherInvocationCalled = false;
-
-    Future<void> cleanup() async {
-      cleanupCalled = true;
-    }
-
-    Future<Future<void> Function()> interaction(WidgetTester tester) async {
-      interactionCalled = true;
-      return cleanup;
-    }
-
-    FutureOr<void> matcherInvocation() {
-      matcherInvocationCalled = true;
-    }
-
-    MatchesGoldenFileInvocation<void> goldenFileExpectation(
-      Object a,
-      Object b,
-    ) {
-      goldenFileExpectationCalled = true;
-      return matcherInvocation;
-    }
 
     setUp(() {
       adapter = MockAdapter();
       goldenTestAdapter = adapter;
 
-      cleanupCalled = false;
-      interactionCalled = false;
-      goldenFileExpectationCalled = false;
-      matcherInvocationCalled = false;
-
       when(
         () => goldenTestAdapter.pumpGoldenTest(
-          rootKey: any(named: 'rootKey'),
           tester: any(named: 'tester'),
           textScaleFactor: any(named: 'textScaleFactor'),
           constraints: any(named: 'constraints'),
-          theme: any(named: 'theme'),
+          obscureFont: any(named: 'obscureFont'),
+          variantConfigTheme: any(named: 'variantConfigTheme'),
+          globalConfigTheme: any(named: 'globalConfigTheme'),
           pumpBeforeTest: any(named: 'pumpBeforeTest'),
           pumpWidget: any(named: 'pumpWidget'),
           widget: any(named: 'widget'),
@@ -101,7 +72,7 @@ void main() {
 
       when(
         () => goldenTestAdapter.goldenFileExpectation,
-      ).thenReturn(goldenFileExpectation);
+      ).thenReturn((_, __) => () async {});
 
       when(
         () => goldenTestAdapter.withForceUpdateGoldenFiles<void>(
@@ -120,89 +91,10 @@ void main() {
         goldenTestRunner.run(
           tester: tester,
           goldenPath: 1,
-          widget: Container(),
+          widget: const SizedBox(),
         ),
         throwsAssertionError,
       );
-    });
-
-    testWidgets('invokes everything correctly', (tester) async {
-      const themeColor = Colors.blue;
-      final theme = ThemeData.light().copyWith(
-        primaryColor: themeColor,
-      );
-      await goldenTestRunner.run(
-        tester: tester,
-        goldenPath: 'path/to/golden',
-        widget: Container(),
-        theme: theme,
-        whilePerforming: interaction,
-        obscureText: true,
-      );
-
-      expect(interactionCalled, isTrue);
-      expect(cleanupCalled, isTrue);
-      expect(goldenFileExpectationCalled, isTrue);
-      expect(matcherInvocationCalled, isTrue);
-
-      final capturedTheme = verify(
-        () => adapter.pumpGoldenTest(
-          rootKey: any(named: 'rootKey'),
-          tester: any(named: 'tester'),
-          textScaleFactor: any(named: 'textScaleFactor'),
-          constraints: any(named: 'constraints'),
-          theme: captureAny(named: 'theme'),
-          pumpBeforeTest: any(named: 'pumpBeforeTest'),
-          pumpWidget: any(named: 'pumpWidget'),
-          widget: any(named: 'widget'),
-        ),
-      ).captured.first as ThemeData;
-
-      expect(
-        capturedTheme,
-        isA<ThemeData>().having(
-          (theme) => theme.primaryColor,
-          'primaryColor',
-          themeColor,
-        ),
-      );
-
-      expect(
-        capturedTheme,
-        isA<ThemeData>().having(
-          (theme) => theme.textTheme.bodyText1!.fontFamily,
-          'textTheme.bodyText1!.fontFamily',
-          obscuredTextFontFamily,
-        ),
-      );
-    });
-
-    testWidgets('invokes everything with defaults', (tester) async {
-      await goldenTestRunner.run(
-        tester: tester,
-        goldenPath: 'path/to/golden',
-        widget: Container(),
-      );
-
-      expect(interactionCalled, isFalse);
-      expect(cleanupCalled, isFalse);
-      expect(goldenFileExpectationCalled, isTrue);
-      expect(matcherInvocationCalled, isTrue);
-
-      final capturedTheme = verify(
-        () => adapter.pumpGoldenTest(
-          rootKey: any(named: 'rootKey'),
-          tester: any(named: 'tester'),
-          textScaleFactor: any(named: 'textScaleFactor'),
-          constraints: any(named: 'constraints'),
-          theme: captureAny(named: 'theme'),
-          pumpBeforeTest: any(named: 'pumpBeforeTest'),
-          pumpWidget: any(named: 'pumpWidget'),
-          widget: any(named: 'widget'),
-        ),
-      ).captured.first as ThemeData;
-
-      expect(capturedTheme, equals(ThemeData.light()));
     });
 
     testWidgets('throws when matcher fails', (tester) async {
@@ -226,7 +118,7 @@ void main() {
         await goldenTestRunner.run(
           tester: tester,
           goldenPath: 'path/to/golden',
-          widget: Container(),
+          widget: const SizedBox(),
         );
         fail('Expected goldenTestRunner.run to throw TestFailure');
       } catch (e) {
@@ -245,7 +137,7 @@ void main() {
           tester: tester,
           goldenPath: 'path/to/golden',
           renderShadows: true,
-          widget: Container(),
+          widget: const SizedBox(),
           whilePerforming: (_) {
             debugDisableShadowsDuringTestRun = debugDisableShadows;
             throw givenException;
@@ -256,6 +148,44 @@ void main() {
 
       expect(debugDisableShadows, isTrue);
       expect(debugDisableShadowsDuringTestRun, isFalse);
+    });
+
+    testWidgets('resets window size after the test has run', (tester) async {
+      late final Size sizeDuringTestRun;
+      final originalSize = tester.binding.window.physicalSize;
+      when(
+        () => goldenTestAdapter.pumpGoldenTest(
+          tester: any(named: 'tester'),
+          textScaleFactor: any(named: 'textScaleFactor'),
+          constraints: any(named: 'constraints'),
+          pumpBeforeTest: any(named: 'pumpBeforeTest'),
+          pumpWidget: any(named: 'pumpWidget'),
+          widget: any(named: 'widget'),
+          obscureFont: any(named: 'obscureFont'),
+          globalConfigTheme: any(named: 'globalConfigTheme'),
+          variantConfigTheme: any(named: 'variantConfigTheme'),
+        ),
+      ).thenAnswer((_) async {
+        tester.binding.window.physicalSizeTestValue = Size.zero;
+      });
+
+      final givenException = Exception();
+      await expectLater(
+        goldenTestRunner.run(
+          tester: tester,
+          goldenPath: 'path/to/golden',
+          renderShadows: true,
+          widget: const SizedBox.square(dimension: 200),
+          whilePerforming: (testerDuringTestRun) {
+            sizeDuringTestRun = testerDuringTestRun.binding.window.physicalSize;
+            throw givenException;
+          },
+        ),
+        throwsA(same(givenException)),
+      );
+
+      expect(tester.binding.window.physicalSize, originalSize);
+      expect(sizeDuringTestRun, Size.zero);
     });
 
     tearDownAll(() {